+++ lib/talkjs.dart
@@ -1,4 +1,4 @@
 library talk_js;
 
 export 'src/talk_js_base.dart';
-export 'src/talk_js_interop.dart';
+export 'src/interop.dart' hide ready, oneOnOneId, getAppMetadata;